+++ decidim-core/app/forms/decidim/editor_image_form.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Decidim
+  class EditorImageForm < Decidim::Form
+    mimic :editor_image
+
+    attribute :file
+    attribute :author_id, Integer
+
+    validates :author_id, presence: true
+    validates :file, presence: true
+    validates :file, passthru: { to: Decidim::EditorImage }
+
+    alias organization current_organization
+  end
+end