+++ decidim-core/db/seeds.rb
@@ -132,21 +132,23 @@ if !Rails.env.production? || ENV["SEED"]
     admin_terms_accepted_at: Time.current
   )
 
-  regular_user = Decidim::User.find_or_initialize_by(email: "user@example.org")
+  ["user@example.org", "user2@example.org"].each do |email|
+    Decidim::User.find_or_initialize_by(email: email).update!(
+      name: Faker::Name.name,
+      nickname: Faker::Twitter.unique.screen_name,
+      password: "decidim123456",
+      password_confirmation: "decidim123456",
+      confirmed_at: Time.current,
+      locale: I18n.default_locale,
+      organization: organization,
+      tos_agreement: true,
+      personal_url: Faker::Internet.url,
+      about: Faker::Lorem.paragraph(sentence_count: 2),
+      accepted_tos_version: organization.tos_version
+    )
+  end
 
-  regular_user.update!(
-    name: Faker::Name.name,
-    nickname: Faker::Twitter.unique.screen_name,
-    password: "decidim123456",
-    password_confirmation: "decidim123456",
-    confirmed_at: Time.current,
-    locale: I18n.default_locale,
-    organization: organization,
-    tos_agreement: true,
-    personal_url: Faker::Internet.url,
-    about: Faker::Lorem.paragraph(sentence_count: 2),
-    accepted_tos_version: organization.tos_version
-  )
+  regular_user = Decidim::User.find_or_initialize_by(email: "user@example.org")
 
   locked_user = Decidim::User.find_or_initialize_by(email: "locked_user@example.org")