+++ decidim-meetings/app/permissions/decidim/meetings/permissions.rb
@@ -91,7 +91,6 @@ module Decidim
       def can_close_meeting?
         component_settings&.creation_enabled_for_participants? &&
           meeting.authored_by?(user) &&
-          !meeting.closed? &&
           meeting.past?
       end