+++ decidim-meetings/spec/spec_helper.rb
@@ -9,3 +9,4 @@ Decidim::Dev.dummy_app_path = File.expand_path(File.join("..", "spec", "decidim_
 require "decidim/dev/test/base_spec_helper"
 
 require "decidim/forms/test"
+require "decidim/comments/test"